feat : Name 객체 복구
